feature: Cross-platform DuckDB facade + JSON schema inference on Node/Native
Lays the groundwork for the unified DuckDB wrapper Leo asked about:

  DuckDB (shared object)        ──▶ DuckDBCompat (per-platform impl)
  DuckDBAnalyzer (shared)       ──▶ dispatches by file extension

`DuckDBCompat` mirrors the SourceIOCompat pattern. JVM keeps the JDBC
backend; JS and Native ship explicit `UnsupportedOperationException`
stubs so a future PR can plug in `@duckdb/node-api` (Node) and
`libduckdb` C bindings (Native) without touching call sites.

Immediate behavioral win: `DuckDBAnalyzer.guessSchema` now routes
`*.json` and `*.json.gz` through `JSONAnalyzer` (which is fully
cross-platform — uni.io for read + uni.json for parse). That means
`from 'foo.json'` queries compile on Node now, where they previously
threw `UnsupportedOperationException` because the only schema path
went through DuckDB JDBC.

Verified end-to-end: the Node CLI (`node main.js compile -f
spec/basic/select-as.wv`) produces valid SQL — was throwing on the
unmerged main of #1694.

Tests: JVM 1399, JS 1379, Native 1378 — all unchanged.

Out of scope (clean follow-ups):
  - real `DuckDB.schemaOf` on Node via `@duckdb/node-api`
  - real `DuckDB.schemaOf` on Native via libduckdb extern bindings
  - growing the facade to back actual query execution on Node/Native
