--- wvlet-lang/.js/src/main/scala/wvlet/lang/compiler/analyzer/DuckDBSchemaAnalyzer.scala
@@ -1,14 +1,0 @@
-package wvlet.lang.compiler.analyzer
-
-import wvlet.lang.model.RelationType
-
-object DuckDBAnalyzer extends DuckDBSchemaAnalyzerCompat:
-  def guessSchema(path: String): RelationType = analyzeFileSchema(path)
-
-trait DuckDBSchemaAnalyzerCompat:
-  protected def analyzeFileSchema(path: String): RelationType =
-    throw new UnsupportedOperationException(
-      "File schema analysis is not supported in Scala.js (browser environment)"
-    )
-
-end DuckDBSchemaAnalyzerCompat
